Improve: Refactor the general code a bit, to make repacking easier.
