+++ NexusMods.Archives.Nx.Tests/Tests/Packing/PackingTests.cs
@@ -3,8 +3,8 @@ using AutoFixture.Xunit2;
 using FluentAssertions;
 using NexusMods.Archives.Nx.Enums;
 using NexusMods.Archives.Nx.FileProviders;
-using NexusMods.Archives.Nx.Headers.Managed;
-using NexusMods.Archives.Nx.Packing;
+using NexusMods.Archives.Nx.Packing.Pack;
+using NexusMods.Archives.Nx.Packing.Unpack;
 using NexusMods.Archives.Nx.Structs;
 using NexusMods.Archives.Nx.Tests.Utilities;
 using NexusMods.Hashing.xxHash64;
